refactor tests into separate block and file for drizzle options
