+++ test/drizzle/constants.js
@@ -1,0 +1,3 @@
+export const NETWORK_MAINNET = 1
+export const NETWORK_RINKEBY = 4
+export const NETWORK_GANACHE = global.defaultNetworkId
